requirements.txt
astroid==1.6.4
Django==2.0.8
django-always-authenticated==0.1.0
django-cors-headers==2.2.0
djangorestframework==3.7.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pbr==3.1.1
Pillow==5.1.0
pylint==1.9.3
pytz==2018.3
six==1.11.0
stevedore==1.25.0
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
wrapt==1.10.11
